requirements.txt
aiogram>=3.7.0
python-dotenv>=1.0.1
instagrapi>=2.1.2
aiohttp>=3.10.5
